Revisited all WinMM interop signatures in an attempt to make them safe with x64. (NAudio still builds as x86 only until testing is possible). Nothing seems to have broken in x86 but proceed with caution.
Also, have fixed a longstanding issue with hanging in WaveOutReset due to function callbacks (thanks to Matt69 on the forums)

+++ NAudio/Wave/Compression/AcmFormatChoose.cs
@@ -7,6 +7,7 @@ namespace NAudio.Wave.Compression
 {
     /// <summary>
     /// ACMFORMATCHOOSE
+    /// http://msdn.microsoft.com/en-us/library/dd742911%28VS.85%29.aspx
     /// </summary>
     [StructLayout(LayoutKind.Sequential)]
     struct AcmFormatChoose
@@ -77,7 +78,7 @@ namespace NAudio.Wave.Compression
         /// <summary>
         /// LPARAM lCustData; 
         /// </summary>
-        public int customData;
+        public IntPtr customData;
         /// <summary>
         /// ACMFORMATCHOOSEHOOKPROC pfnHook; 
         /// </summary>
